ex05 done

+++ ex05/ft_sqrt.c
@@ -6,15 +6,15 @@
 /*   By: yrabby <yrabby@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/25 12:59:30 by yrabby            #+#    #+#             */
-/*   Updated: 2022/05/25 13:44:16 by yrabby           ###   ########.fr       */
+/*   Updated: 2022/05/25 14:34:30 by yrabby           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 int	ft_sqrt(int nb)
 {
-	int	start;
-	int	end;
-	int	mid;
+	long	start;
+	long	end;
+	long	mid;
 
 	start = 0;
 	end = nb;
@@ -24,7 +24,7 @@ int	ft_sqrt(int nb)
 	{
 		mid = (start + end) / 2;
 		if (mid * mid == nb)
-			return (mid);
+			return ((int)mid);
 		if (mid * mid < nb)
 			start = mid + 1;
 		else
